some changes in lab-2
